centralize Rust integration and regression tests
