--- test/README.md
@@ -1,5 +1,0 @@
-Currently the best way to run these tests are like this due to long running tests:
-
-```bash
-cargo test test_sequential_write -- --nocapture
-```
