+++ tests/Cargo.toml
@@ -5,12 +5,14 @@ authors.workspace = true
 edition.workspace = true
 license.workspace = true
 repository.workspace = true
-description = "Internal tester of write path"
+description = "Integration tests"
 
 [lib]
-name = "test"
-path = "src/lib.rs"
+path = "lib.rs"
 
+[[test]]
+name = "integration_tests"
+path = "integration/mod.rs"
 
 [dependencies]
 anyhow = "1.0.75"
@@ -22,6 +24,8 @@ rustyline = "12.0.0"
 rusqlite = { version = "0.29", features = ["bundled"] }
 tempfile = "3.0.7"
 log = "0.4.22"
+assert_cmd = "^2"
 
-[dev-dependencies]
-rstest = "0.18.2"
+# rexpect does not support windows.
+[target.'cfg(not(windows))'.dependencies]
+rexpect = "0.6.0"
