+++ tests/README.md
@@ -1,0 +1,11 @@
+
+Integration and regression test suite.
+
+```bash
+
+# run all tests
+cargo test 
+
+# run individual test
+cargo test test_sequential_write -- --nocapture
+```
