+++ tests/integration/mod.rs
@@ -1,0 +1,4 @@
+mod common;
+mod functions;
+mod pragma;
+mod query_processing;
